server/requirements.txt
alembic==1.14.1
bcrypt==5.0.0
blinker==1.8.2
certifi
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
cryptography==46.0.7
exceptiongroup==1.3.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
gunicorn==23.0.0
idna==3.11
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
packaging
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.0.1
python-http-client==3.3.7
requests==2.32.4
resend==2.28.0
sendgrid==6.12.5
SQLAlchemy==2.0.48
tomli==2.4.1
typing_extensions==4.13.2
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
